% Auto-generated: do not edit by hand
\name{accordionItem}

\alias{accordionItem}

\title{AccordionItem component}

\description{
Utility component to pass to Accordion. For more information, see: https://mantine.dev/core/accordion/
}

\usage{
accordionItem(children=NULL, description=NULL, label=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Accordion content}

\item{description}{Character. Accordion description}

\item{label}{Character. Accordion label}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

