% Auto-generated: do not edit by hand
\name{affix}

\alias{affix}

\title{Affix component}

\description{
Renders Affix at given element on mouse over or any other event. For more information, see: https://mantine.dev/core/affix/
}

\usage{
affix(children=NULL, id=NULL, className=NULL, position=NULL,
style=NULL, zIndex=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Any react node that should trigger tooltip}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{position}{Lists containing elements 'top', 'left', 'bottom', 'right'.
those elements have the following types:
  - top (numeric | character; optional)
  - left (numeric | character; optional)
  - bottom (numeric | character; optional)
  - right (numeric | character; optional). Fixed position in px}

\item{style}{Named list. Inline style override}

\item{zIndex}{Character. Root element z-index property}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

