% Auto-generated: do not edit by hand
\name{anchor}

\alias{anchor}

\title{Anchor component}

\description{
Display links with theme styles. For more information, see: https://mantine.dev/core/anchor/
}

\usage{
anchor(children=NULL, id=NULL, align=NULL, className=NULL,
color=NULL, gradient=NULL, href=NULL, inherit=NULL,
inline=NULL, lineClamp=NULL, size=NULL, style=NULL,
target=NULL, transform=NULL, variant=NULL, weight=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Primary content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{align}{Sets text-align css property}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Text color from theme}

\item{gradient}{Controls gradient settings in gradient variant only}

\item{href}{Character. href}

\item{inherit}{Logical. Inherit font properties from parent element}

\item{inline}{Logical. Sets line-height to 1 for centering}

\item{lineClamp}{Numeric. CSS -webkit-line-clamp property}

\item{size}{Predefined font-size from theme.fontSizes}

\item{style}{Named list. Inline style override}

\item{target}{A value equal to: "_blank", "_self". Target}

\item{transform}{Sets text-transform css property}

\item{variant}{Link or text variant}

\item{weight}{Sets font-weight css property}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

