% Auto-generated: do not edit by hand
\name{blockquote}

\alias{blockquote}

\title{Blockquote component}

\description{
Blockquote with optional cite. For more information, see: https://mantine.dev/core/blockquote/
}

\usage{
blockquote(children=NULL, id=NULL, cite=NULL, className=NULL,
color=NULL, style=NULL)
}

\arguments{
\item{children}{Character. Primary content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{cite}{Character. Describe a reference to a cited quote}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Badge color from theme}

\item{style}{Named list. Inline style override}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

