% Auto-generated: do not edit by hand
\name{cardSection}

\alias{cardSection}

\title{CardSection component}

\description{
Card.Section is a special component that is used to remove Card padding from its children while other elements still have horizontal spacing. For more information, see: https://mantine.dev/core/card/
}

\usage{
cardSection(children=NULL, id=NULL, className=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Section content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

