% Auto-generated: do not edit by hand
\name{center}

\alias{center}

\title{Center component}

\description{
Centers content vertically and horizontally. For more information, see: https://mantine.dev/core/center/
}

\usage{
center(children=NULL, id=NULL, class_name=NULL, inline=NULL,
style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content that should be centered vertically and horizontally}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{inline}{Logical. Set to true to use inline-flex instead of flex}

\item{style}{Named list. Inline style override}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

