% Auto-generated: do not edit by hand
\name{checkbox}

\alias{checkbox}

\title{Checkbox component}

\description{
Capture boolean input from user. For more information, see: https://mantine.dev/core/checkbox/
}

\usage{
checkbox(id=NULL, checked=NULL, class_name=NULL, color=NULL,
disabled=NULL, label=NULL, size=NULL, style=NULL,
transitionDuration=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{checked}{Logical. State of check box}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{color}{A value equal to: "dark", "gray", "red", "pink", "grape", "violet", "indigo", "blue", "cyan", "teal", "green", "lime", "yellow", "orange". Checkbox color}

\item{disabled}{Logical. A checkbox can show it is currently unable to be interacted with}

\item{label}{Character. Checkbox label}

\item{size}{A value equal to: "xs", "sm", "md", "lg", "xl". Predefined label font-size and checkbox width and height in px}

\item{style}{Named list. Inline style override}

\item{transitionDuration}{Numeric. Check/uncheck transition duration, set to 0 to disable all transitions}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

