% Auto-generated: do not edit by hand
\name{chips}

\alias{chips}

\title{Chips component}

\description{
Alternative to Select and RadioGroup. For more information, see: https://mantine.dev/core/chips/
}

\usage{
chips(id=NULL, align=NULL, className=NULL, color=NULL, data=NULL,
direction=NULL, grow=NULL, multiple=NULL, noWrap=NULL,
position=NULL, radius=NULL, size=NULL, spacing=NULL,
style=NULL, value=NULL, variant=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{align}{Defines align-items css property}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Active chip color, defaults to theme.primaryColor}

\item{data}{Chips}

\item{direction}{Defines flex-direction property, row for horizontal, column for vertical}

\item{grow}{Logical. Defines flex-grow property for each element, true -> 1, false -> 0}

\item{multiple}{Logical. Allow multiple values to be picked}

\item{noWrap}{Logical. Defined flex-wrap property}

\item{position}{Defines justify-content property}

\item{radius}{Chip border-radius from theme or number to set value in px}

\item{size}{Predefined chip size}

\item{spacing}{Spacing between chips from theme or number to set value in px}

\item{style}{Named list. Inline style override}

\item{value}{Character | list of characters. Controlled component value}

\item{variant}{A value equal to: "outline", "filled". Controls chip appearance, defaults to filled with dark theme and to outline in light theme}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

