% Auto-generated: do not edit by hand
\name{colorPicker}

\alias{colorPicker}

\title{ColorPicker component}

\description{
Alternative to Select and RadioGroup. For more information, see: https://mantine.dev/core/ColorPicker/
}

\usage{
colorPicker(id=NULL, alphaLabel=NULL, class_name=NULL, focusable=NULL,
format=NULL, fullWidth=NULL, hueLabel=NULL,
saturationLabel=NULL, size=NULL, style=NULL, swatches=NULL,
swatchesPerRow=NULL, value=NULL, withPicker=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{alphaLabel}{Character. Alpha slider aria-label}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{focusable}{Logical. Should interactive elements be focusable}

\item{format}{A value equal to: "hex", "rgba", "rgb", "hsl", "hsla". Color format}

\item{fullWidth}{Logical. Force picker to take 100\% width of its container}

\item{hueLabel}{Character. Hue slider aria-label}

\item{saturationLabel}{Character. Saturation slider aria-label}

\item{size}{A value equal to: "xs", "sm", "md", "lg", "xl". Predefined component size}

\item{style}{Named list. Inline style override}

\item{swatches}{List of characters. Predefined colors}

\item{swatchesPerRow}{Numeric. Number of swatches displayed in one row}

\item{value}{Character. Controlled component value}

\item{withPicker}{Logical. Set to false to display swatches only}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

