% Auto-generated: do not edit by hand
\name{container}

\alias{container}

\title{Container component}

\description{
Center content horizontally with predefined max-width. For more information, see: https://mantine.dev/core/container/
}

\usage{
container(children=NULL, id=NULL, className=NULL, fluid=NULL,
padding=NULL, size=NULL, style=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Content that should be centered vertically and horizontally}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{fluid}{Logical. If fluid is set to true, size prop is ignored and Container always take 100\% of width}

\item{padding}{Horizontal padding defined in theme.spacing, or number value for padding in px}

\item{size}{Predefined container max-width or number for max-width in px}

\item{style}{Named list. Inline style override}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

