% Auto-generated: do not edit by hand
\name{notification}

\alias{notification}

\title{Notification component}

\description{
Show dynamic notifications and alerts to user, part of notifications system. For more information, see: https://mantine.dev/core/notification/
}

\usage{
notification(children=NULL, id=NULL, className=NULL, color=NULL,
loading=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Notification body, place main text here}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Notification line or icon color}

\item{loading}{Logical. Replaces colored line or icon with Loader component}

\item{title}{Character. Notification title, displayed before body}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

