% Auto-generated: do not edit by hand
\name{scrollArea}

\alias{scrollArea}

\title{ScrollArea component}

\description{
A port of the ScrollArea component. For more information, see: https://mantine.dev/core/table/
}

\usage{
scrollArea(children=NULL, id=NULL, className=NULL, dir=NULL,
offsetScrollbars=NULL, scrollHideDelay=NULL,
scrollbarSize=NULL, style=NULL, type=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. ScrollArea children}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{dir}{A value equal to: "ltr", "rtl". Reading direction of the scroll area}

\item{offsetScrollbars}{Logical. Should scrollbars be offset with padding}

\item{scrollHideDelay}{Numeric. Scroll hide delay in ms, for scroll and hover types only}

\item{scrollbarSize}{Numeric. Scrollbar size in px}

\item{style}{Named list. Inline style override}

\item{type}{A value equal to: "auto", "scroll", "always", "hover". Scrollbars type}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

