% Auto-generated: do not edit by hand
\name{skeleton}

\alias{skeleton}

\title{Skeleton component}

\description{
Indicate content loading state. For more information, see: https://mantine.dev/core/skeleton/
}

\usage{
skeleton(children=NULL, id=NULL, circle=NULL, height=NULL,
radius=NULL, visible=NULL, width=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Primary content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{circle}{Logical. If Skeleton is a circle, it's width and border-radius will be equal to height}

\item{height}{Skeleton height}

\item{radius}{Radius from theme.radius or number to set border-radius in px}

\item{visible}{Logical. Should skeleton overlay be displayed}

\item{width}{Skeleton width}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

