% Auto-generated: do not edit by hand
\name{tab}

\alias{tab}

\title{Tab component}

\description{
Utility component to pass to Tabs. For more information, see: https://mantine.dev/core/tabs/
}

\usage{
tab(children=NULL, id=NULL, className=NULL, disabled=NULL,
label=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Tab content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{disabled}{Logical. A tab can show it is currently unable to be interacted with}

\item{label}{Character. Tab control label}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

