% Auto-generated: do not edit by hand
\name{tabs}

\alias{tabs}

\title{Tabs component}

\description{
Switch between different views. For more information, see: https://mantine.dev/core/tabs/
}

\usage{
tabs(children=NULL, id=NULL, active=NULL, className=NULL,
color=NULL, grow=NULL, orientation=NULL, position=NULL,
tabPadding=NULL, variant=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. <Tab /> components only}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{active}{Numeric. Index of active tab, overrides internal state}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Active tab color from theme.colors}

\item{grow}{Logical. True if tabs should take all available space}

\item{orientation}{Controls tab orientation}

\item{position}{Tab controls position}

\item{tabPadding}{Controls tab content padding-top}

\item{variant}{A value equal to: "default", "outline", "pills". Controls appearance}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

