% Auto-generated: do not edit by hand
\name{text}

\alias{text}

\title{Text component}

\description{
Render text and links with theme styles. For more information, see: https://mantine.dev/core/text/
}

\usage{
text(children=NULL, id=NULL, align=NULL, className=NULL,
color=NULL, gradient=NULL, inherit=NULL, inline=NULL,
lineClamp=NULL, radius=NULL, size=NULL, style=NULL,
transform=NULL, variant=NULL, weight=NULL)
}

\arguments{
\item{children}{Character. Primary content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{align}{A value equal to: "left", "right", "center". Sets text-align css property}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Text color from theme}

\item{gradient}{Controls gradient settings in gradient variant only}

\item{inherit}{Logical. Inherit font properties from parent element}

\item{inline}{Logical. Sets line-height to 1 for centering}

\item{lineClamp}{Numeric. CSS -webkit-line-clamp property}

\item{radius}{Button border-radius from theme or number to set border-radius in px}

\item{size}{Predefined font-size from theme.fontSizes}

\item{style}{Named list. Inline style override}

\item{transform}{A value equal to: "capitalize", "uppercase", "lowercase". Sets text-transform css property}

\item{variant}{A value equal to: "link", "gradient", "text". Link or text variant}

\item{weight}{A value equal to: "normal", "bold", "bolder", "lighter", "initial", "inherit" | numeric. Sets font-weight css property}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

