% Auto-generated: do not edit by hand
\name{accordionItem}

\alias{accordionItem}

\title{AccordionItem component}

\description{
Utility component to pass to Accordion. For more information, see: https://mantine.dev/core/accordion/
}

\usage{
accordionItem(children=NULL, id=NULL, className=NULL, description=NULL,
label=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Col content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{description}{Character. Accordion description}

\item{label}{Character. Accordion label}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

