% Auto-generated: do not edit by hand
\name{alert}

\alias{alert}

\title{Alert component}

\description{
Attract user attention with important static message. For more information, see: https://mantine.dev/core/alert/
}

\usage{
alert(children=NULL, id=NULL, className=NULL, color=NULL,
style=NULL, title=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Alert message}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Alert title and line colors from theme}

\item{style}{Named list. Inline style override}

\item{title}{Character. Optional alert title}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

