% Auto-generated: do not edit by hand
\name{breadcrumbs}

\alias{breadcrumbs}

\title{Breadcrumbs component}

\description{
Separate list of react nodes with given separator. For more information, see: https://mantine.dev/core/breadcrumbs/
}

\usage{
breadcrumbs(id=NULL, class_name=NULL, items=NULL, separator=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{items}{List of lists containing elements 'title', 'href'.
those elements have the following types:
  - title (character; required)
  - href (character; required)s. Link items}

\item{separator}{Character. Separator between breadcrumbs}

\item{style}{Named list. Inline style override}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

