% Auto-generated: do not edit by hand
\name{space}

\alias{space}

\title{Space component}

\description{
Add horizontal or vertical spacing from theme. For more information, see: https://mantine.dev/core/space/
}

\usage{
space(children=NULL, id=NULL, class_name=NULL, h=NULL, style=NULL,
w=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. Tab content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{h}{A value equal to: "xs", "sm", "md", "lg", "xl" | numeric. Height, set to add vertical spacing}

\item{style}{Named list. Inline style override}

\item{w}{A value equal to: "xs", "sm", "md", "lg", "xl" | numeric. Width, set to add horizontal spacing}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

