% Auto-generated: do not edit by hand
\name{switch}

\alias{switch}

\title{Switch component}

\description{
Capture user feedback limited to small set of options. For more information, see: https://mantine.dev/core/switch/
}

\usage{
switch(id=NULL, checked=NULL, className=NULL, color=NULL,
disabled=NULL, label=NULL, radius=NULL, size=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{checked}{Logical. State of check box}

\item{className}{Character. Often used with CSS to style elements with common properties}

\item{color}{Checkbox color}

\item{disabled}{Logical. A checkbox can show it is currently unable to be interacted with}

\item{label}{Character. Checkbox label}

\item{radius}{Predefined border-radius value from theme.radius or number for border-radius in px}

\item{size}{Predefined label font-size and checkbox width and height in px}

\item{style}{Named list. Inline style override}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

