% Auto-generated: do not edit by hand
\name{tabs}

\alias{tabs}

\title{Tabs component}

\description{
Switch between different views. For more information, see: https://mantine.dev/core/tabs/
}

\usage{
tabs(children=NULL, id=NULL, active=NULL, class_name=NULL,
color=NULL, grow=NULL, orientation=NULL, position=NULL,
style=NULL, tabPadding=NULL, variant=NULL)
}

\arguments{
\item{children}{A list of or a singular dash component, string or number. <Tab /> components only}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{active}{Numeric. Index of active tab, overrides internal state}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{color}{A value equal to: "dark", "gray", "red", "pink", "grape", "violet", "indigo", "blue", "cyan", "teal", "green", "lime", "yellow", "orange". Active tab color from theme.colors}

\item{grow}{Logical. True if tabs should take all available space}

\item{orientation}{A value equal to: "horizontal", "vertical". Controls tab orientation}

\item{position}{A value equal to: "right", "center", "left", "apart". Tab controls position}

\item{style}{Named list. Inline style override}

\item{tabPadding}{A value equal to: "xs", "sm", "md", "lg", "xl". Controls tab content padding-top}

\item{variant}{A value equal to: "default", "outline", "pills". Controls appearance}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

