% Auto-generated: do not edit by hand
\name{title}

\alias{title}

\title{Title component}

\description{
h1-h6 headings. For more information, see: https://mantine.dev/core/title/
}

\usage{
title(children=NULL, id=NULL, align=NULL, class_name=NULL,
order=NULL, style=NULL)
}

\arguments{
\item{children}{Character. Primary content}

\item{id}{Character. The ID of this component, used to identify dash components in callbacks}

\item{align}{A value equal to: "left", "right", "center". Sets text-align css property}

\item{class_name}{Character. Often used with CSS to style elements with common properties}

\item{order}{Numeric. Defines component and styles which will be used}

\item{style}{Named list. Inline style override}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

